+++ src/test/groovy/io/codearte/gradle/nexus/functional/MockedFunctionalSpec.groovy
@@ -4,11 +4,10 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule
 import com.github.tomakehurst.wiremock.stubbing.Scenario
 import groovy.json.JsonOutput
 import io.codearte.gradle.nexus.logic.FetcherResponseTrait
+import io.codearte.gradle.nexus.logic.RepositoryState
 import nebula.test.functional.ExecutionResult
 import org.gradle.api.logging.LogLevel
 import org.junit.Rule
-import spock.lang.Ignore
-import spock.lang.Issue
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse
 import static com.github.tomakehurst.wiremock.client.WireMock.containing
@@ -20,17 +19,22 @@ import static com.github.tomakehurst.wiremock.client.WireMock.stubFor
 import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo
 import static com.github.tomakehurst.wiremock.client.WireMock.verify
 
+//TODO: Split into two files: basic and transition related
 class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements FetcherResponseTrait {
 
     @Rule
     public WireMockRule wireMockRule = new WireMockRule(8089)
 
-    protected static final String stagingProfileId = "93c08fdebde1ff"
+    private static final String stagingProfileId = "5027d084a01a3a"
+    private static final String REPO_ID_1 = "testRepo1"
+    private static final String REPO_ID_2 = "testRepo2"
 
     def "should not do request for staging profile when provided in configuration on #testedTaskName task"() {
         given:
             stubGetOneRepositoryWithProfileIdAndContent(stagingProfileId,
-                    createResponseMapWithGivenRepos([aRepoInStateAndId(repoTypeToReturn, "ignored")]))
+                    createResponseMapWithGivenRepos([aRepoInStateAndId(REPO_ID_1, repoTypeToReturn)]))
+        and:
+            stubGetRepositoryStateByIdForConsecutiveStates(REPO_ID_1, repositoryStatesToGetById)
         and:
             stubSuccessfulCloseRepositoryWithProfileId(stagingProfileId)
         and:
@@ -52,9 +56,9 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
         and:
             verify(0, getRequestedFor(urlEqualTo("/staging/profiles")))
         where:
-            testedTaskName      | repoTypeToReturn
-            "closeRepository"   | "open"
-            "promoteRepository" | "closed"
+            testedTaskName      | repoTypeToReturn       | repositoryStatesToGetById
+            "closeRepository"   | RepositoryState.OPEN   | [RepositoryState.CLOSED]
+            "promoteRepository" | RepositoryState.CLOSED | [RepositoryState.RELEASED]
     }
 
     def "should send request for staging profile when not provided in configuration on #testedTaskName task"() {
@@ -62,7 +66,9 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
             stubGetStagingProfilesWithJson(this.getClass().getResource("/io/codearte/gradle/nexus/logic/2stagingProfilesShrunkResponse.json").text)
         and:
             stubGetOneRepositoryWithProfileIdAndContent(stagingProfileId,
-                    createResponseMapWithGivenRepos([aRepoInStateAndId(repoTypeToReturn, "ignored")]))
+                    createResponseMapWithGivenRepos([aRepoInStateAndId(REPO_ID_1, repoTypeToReturn)]))
+        and:
+            stubGetRepositoryStateByIdForConsecutiveStates(REPO_ID_1, repositoryStatesToGetById)
         and:
             stubSuccessfulCloseRepositoryWithProfileId(stagingProfileId)
         and:
@@ -71,9 +77,12 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
             buildFile << """
                 ${getApplyPluginBlock()}
                 ${getDefaultConfigurationClosure()}
+                nexusStaging {
+                    stagingProfileId = null //by default set in tests
+                }
             """.stripIndent()
         when:
-            def result = runTasksSuccessfully(testedTaskName)
+            ExecutionResult result = runTasksSuccessfully(testedTaskName)
         then:
             result.wasExecuted(testedTaskName)
             !result.standardOutput.contains("Using configured staging profile id: $stagingProfileId")
@@ -81,43 +90,19 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
         and:
             verify(1, getRequestedFor(urlEqualTo("/staging/profiles")))
         where:
-            testedTaskName      | repoTypeToReturn
-            "closeRepository"   | "open"
-            "promoteRepository" | "closed"
+            testedTaskName      | repoTypeToReturn       | repositoryStatesToGetById
+            "closeRepository"   | RepositoryState.OPEN   | [RepositoryState.CLOSED]
+            "promoteRepository" | RepositoryState.CLOSED | [RepositoryState.RELEASED]
     }
 
-    //TODO: Remove when 'stagingProfileId AND stagingRepositoryId' case is fixed
-    def "should reuse stagingProfileId from closeRepository in promoteRepository when called together"() {
-        given:
-            stubGetStagingProfilesWithJson(this.getClass().getResource("/io/codearte/gradle/nexus/logic/2stagingProfilesShrunkResponse.json").text)
-        and:
-            stubGetOneOpenRepositoryInFirstCallAndOneClosedInTheNext(stagingProfileId)
-        and:
-            stubSuccessfulCloseRepositoryWithProfileId(stagingProfileId)
-        and:
-            stubSuccessfulPromoteRepositoryWithProfileId(stagingProfileId)
-        and:
-            buildFile << """
-                ${getApplyPluginBlock()}
-                ${getDefaultConfigurationClosure()}
-            """.stripIndent()
-        when:
-            ExecutionResult result = runTasksSuccessfully("closeRepository", "promoteRepository")
-        then:
-            result.wasExecuted("closeRepository")
-            result.wasExecuted("promoteRepository")
-        and:
-            verify(2, getRequestedFor(urlEqualTo("/staging/profile_repositories/$stagingProfileId")))
-            verify(1, getRequestedFor(urlEqualTo("/staging/profiles")))
-    }
-
-    @Ignore("Due to https://github.com/Codearte/gradle-nexus-staging-plugin/issues/44")
     def "should reuse stagingProfileId AND stagingRepositoryId from closeRepository in promoteRepository when called together"() {
         given:
             stubGetStagingProfilesWithJson(this.getClass().getResource("/io/codearte/gradle/nexus/logic/2stagingProfilesShrunkResponse.json").text)
         and:
             stubGetOneOpenRepositoryAndOneClosedInFirstCallAndTwoClosedInTheNext(stagingProfileId)
         and:
+            stubGetRepositoryStateByIdForConsecutiveStates(REPO_ID_1, [RepositoryState.CLOSED, RepositoryState.RELEASED])
+        and:
             stubSuccessfulCloseRepositoryWithProfileId(stagingProfileId)
         and:
             stubSuccessfulPromoteRepositoryWithProfileId(stagingProfileId)
@@ -125,6 +110,9 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
             buildFile << """
                 ${getApplyPluginBlock()}
                 ${getDefaultConfigurationClosure()}
+                nexusStaging {
+                    stagingProfileId = null //by default set in tests
+                }
             """.stripIndent()
         when:
             ExecutionResult result = runTasksSuccessfully("closeRepository", "promoteRepository")
@@ -155,6 +143,8 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
         given:
             stubGetOneOpenRepositoryInFirstCallAndOneClosedInTheNext(stagingProfileId)
         and:
+            stubGetRepositoryStateByIdForConsecutiveStates(REPO_ID_1, [RepositoryState.RELEASED])
+        and:
             stubSuccessfulCloseRepositoryWithProfileId(stagingProfileId)
         and:
             stubSuccessfulPromoteRepositoryWithProfileId(stagingProfileId)
@@ -162,11 +152,6 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
             buildFile << """
                 ${getApplyPluginBlock()}
                 ${getDefaultConfigurationClosure()}
-                nexusStaging {
-                    stagingProfileId = "$stagingProfileId"
-                    delayBetweenRetriesInMillis = 50
-                    numberOfRetries = 2
-                }
             """.stripIndent()
         when:
             ExecutionResult result = runTasksSuccessfully("promoteRepository")
@@ -189,13 +174,14 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
         when:
             ExecutionResult result = runTasksSuccessfully('getStagingProfile')
         then:
-            result.standardOutput.contains("Received staging profile id: 93c08fdebde1ff")
+            result.standardOutput.contains("Received staging profile id: $stagingProfileId")
     }
 
     def "should call close and promote in closeAndPromoteRepository task"() {
         given:
-//            stubGetOneOpenRepositoryAndOneClosedInFirstCallAndTwoClosedInTheNext(stagingProfileId)    //TODO: Temporary disabled due to #44
-            stubGetOneOpenRepositoryInFirstCallAndOneClosedInTheNext(stagingProfileId)
+            stubGetOneOpenRepositoryAndOneClosedInFirstCallAndTwoClosedInTheNext(stagingProfileId)
+        and:
+            stubGetRepositoryStateByIdForConsecutiveStates(REPO_ID_1, [RepositoryState.CLOSED, RepositoryState.RELEASED])
         and:
             stubSuccessfulCloseRepositoryWithProfileId(stagingProfileId)
         and:
@@ -204,9 +190,6 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
             buildFile << """
                 ${getApplyPluginBlock()}
                 ${getDefaultConfigurationClosure()}
-                nexusStaging {
-                    stagingProfileId = "$stagingProfileId"
-                }
             """.stripIndent()
         when:
             ExecutionResult result = runTasksSuccessfully("closeAndPromoteRepository")
@@ -222,15 +205,16 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
         and:
             buildFile << """
                 ${getApplyPluginBlock()}
+                ${getDefaultConfigurationClosure()}
                 nexusStaging {
-                    serverUrl = "http://localhost:8089/"
+                    stagingProfileId = null
                 }
                 project.group = "io.codearte"
             """.stripIndent()
         when:
             ExecutionResult result = runTasksSuccessfully('getStagingProfile')
         then:
-            result.standardOutput.contains("Received staging profile id: 93c08fdebde1ff")
+            result.standardOutput.contains("Received staging profile id: $stagingProfileId")
     }
 
     def "explicitly defined packageGroup should override default value"() {
@@ -245,16 +229,45 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
         when:
             ExecutionResult result = runTasksSuccessfully('getStagingProfile')
         then:
-            result.standardOutput.contains("Received staging profile id: 93c08fdebde1ff")
+            result.standardOutput.contains("Received staging profile id: $stagingProfileId")
+    }
+
+    def "should wait on #operationName operation until transitioning is finished"() {
+        given:
+            stubGetOneRepositoryWithProfileIdAndContent(stagingProfileId,
+                createResponseMapWithGivenRepos([aRepoInStateAndId(REPO_ID_1, repoStates[0])]))
+        and:
+            stubGetRepositoryStateByIdForConsecutiveStates(REPO_ID_1, repoStates, [true, false])
+        and:
+            stubbingOperation()
+        and:
+            buildFile << """
+                ${getApplyPluginBlock()}
+                ${getDefaultConfigurationClosure()}
+            """.stripIndent()
+        when:
+            ExecutionResult result = runTasksSuccessfully("${operationName}Repository")
+        then:
+            result.wasExecuted("${operationName}Repository")
+            result.standardOutput.contains("Attempt 1/3 failed.")
+            !result.standardOutput.contains("Attempt 2/3 failed.")
+        where:
+            operationName | repoStates                                         | stubbingOperation
+            "close"       | [RepositoryState.OPEN, RepositoryState.CLOSED]     | { stubSuccessfulCloseRepositoryWithProfileId(stagingProfileId) }
+            "promote"     | [RepositoryState.CLOSED, RepositoryState.RELEASED] | { stubSuccessfulPromoteRepositoryWithProfileId(stagingProfileId) }
     }
 
     @Override
     protected String getDefaultConfigurationClosure() {
         return """
                 nexusStaging {
+                    stagingProfileId = "$stagingProfileId"
                     username = "codearte"
                     packageGroup = "io.codearte"
                     serverUrl = "http://localhost:8089/"
+                    //To do not wait too long in case of failure
+                    delayBetweenRetriesInMillis = 50
+                    numberOfRetries = 2
                 }
         """
     }
@@ -264,19 +277,39 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
                 .withHeader("Content-Type", containing("application/json"))
                 .withHeader("Accept", containing("application/json"))
                 .willReturn(aResponse()
-                .withStatus(200)
-                .withHeader("Content-Type", "application/json")
-                .withBody(responseAsJson)))
+                    .withStatus(200)
+                    .withHeader("Content-Type", "application/json")
+                    .withBody(responseAsJson)))
     }
 
+    //TODO: Do not pass stagingProfileId as argument - use constant - only one is used everywhere
     private void stubGetOneRepositoryWithProfileIdAndContent(String stagingProfileId, Map response) {
         stubFor(get(urlEqualTo("/staging/profile_repositories/$stagingProfileId"))
                 .withHeader("Content-Type", containing("application/json"))
                 .withHeader("Accept", containing("application/json"))
                 .willReturn(aResponse()
-                .withStatus(200)
-                .withHeader("Content-Type", "application/json")
-                .withBody(JsonOutput.toJson(response))))
+                    .withStatus(200)
+                    .withHeader("Content-Type", "application/json")
+                    .withBody(JsonOutput.prettyPrint(JsonOutput.toJson(response)))))
+    }
+
+    private void stubGetRepositoryStateByIdForConsecutiveStates(String repoId, List<RepositoryState> repoStates,
+                                                                List<Boolean> isTransitioningList = null) {
+        if (isTransitioningList == null) {
+            isTransitioningList = repoStates.collect { false }
+        }
+        repoStates.eachWithIndex { repoState, index ->
+            stubFor(get(urlEqualTo("/staging/repository/$repoId"))
+                .inScenario("StateById")
+                .whenScenarioStateIs(index == 0 ? Scenario.STARTED : repoStates[index].name())
+                .withHeader("Content-Type", containing("application/json"))
+                .withHeader("Accept", containing("application/json"))
+                .willReturn(aResponse()
+                    .withStatus(200)
+                    .withHeader("Content-Type", "application/json")
+                    .withBody(JsonOutput.prettyPrint(JsonOutput.toJson(aRepoInStateAndId(repoId, repoState, isTransitioningList[index])))))
+                .willSetStateTo(repoStates[index < repoStates.size() - 1 ? index + 1 : index].name()))  //TODO: Simplify/extract...
+        }
     }
 
     private void stubSuccessfulCloseRepositoryWithProfileId(String stagingProfileId) {
@@ -291,21 +324,22 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
         stubFor(post(urlEqualTo("/staging/profiles/$stagingProfileId/$restCommandName"))
                 .withHeader("Content-Type", equalTo("application/json"))
                 .withHeader("Accept", containing("application/json"))
+                //TODO: Content matching
                 .willReturn(aResponse()
-                .withStatus(200)
-                .withHeader("Content-Type", "application/json")))
+                    .withStatus(200)
+                    .withHeader("Content-Type", "application/json")))
     }
 
     private void stubGetOneOpenRepositoryAndOneClosedInFirstCallAndTwoClosedInTheNext(String stagingProfileId) {
         stubGetGivenRepositoriesInFirstAndSecondCall(stagingProfileId,
-                [aRepoInStateAndId("open", "ignored"), aRepoInStateAndId("closed", "ignoredClosed")],
-                [aRepoInStateAndId("closed", "ignored"), aRepoInStateAndId("closed", "ignoredClosed")])
+                [aRepoInStateAndId(REPO_ID_1, RepositoryState.OPEN), aRepoInStateAndId(REPO_ID_2, RepositoryState.CLOSED)],
+                [aRepoInStateAndId(REPO_ID_1, RepositoryState.CLOSED), aRepoInStateAndId(REPO_ID_2, RepositoryState.CLOSED)])
     }
 
     private void stubGetOneOpenRepositoryInFirstCallAndOneClosedInTheNext(String stagingProfileId) {
         stubGetGivenRepositoriesInFirstAndSecondCall(stagingProfileId,
-                [aRepoInStateAndId("open", "ignored")],
-                [aRepoInStateAndId("closed", "ignored")])
+                [aRepoInStateAndId(REPO_ID_1, RepositoryState.OPEN)],
+                [aRepoInStateAndId(REPO_ID_1, RepositoryState.CLOSED)])
     }
 
     private void stubGetGivenRepositoriesInFirstAndSecondCall(String stagingProfileId, List<Map> repositoriesToReturnInFirstCall,
@@ -318,7 +352,7 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
                 .withStatus(200)
                 .withHeader("Content-Type", "application/json")
                 .withBody(
-                    JsonOutput.toJson(createResponseMapWithGivenRepos(repositoriesToReturnInFirstCall))
+                    JsonOutput.prettyPrint(JsonOutput.toJson(createResponseMapWithGivenRepos(repositoriesToReturnInFirstCall)))
                 )
             )
             .willSetStateTo("CLOSED"))
@@ -331,7 +365,7 @@ class MockedFunctionalSpec extends BaseNexusStagingFunctionalSpec implements Fet
                 .withStatus(200)
                 .withHeader("Content-Type", "application/json")
                 .withBody(
-                    JsonOutput.toJson(createResponseMapWithGivenRepos(repositoriesToReturnInSecondCall))
+                    JsonOutput.prettyPrint(JsonOutput.toJson(createResponseMapWithGivenRepos(repositoriesToReturnInSecondCall)))
                 )
             )
         )